Default to reading YAML file
