+++ .gitignore
@@ -1,4 +1,3 @@
 .DS_Store
-balena-data-extractor
+/balena-data-extractor
 .vscode
-cmds.yml
