--- cmds.example.yaml
@@ -1,10 +1,0 @@
-api:
-    name: GitHub API
-    url: https://api.github.com
-    key: your-api-key-or-remove-this-line
-file:
-    name: OS Release Info
-    path: /etc/os-release
-shell:
-    name: Contents of the programme folder
-    cmd: ls -lah
